+++ src/cadqueryhelper/__init__.py
@@ -1,1 +1,2 @@
 from .series import series
+from .irregularGrid import irregular_grid